% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methScores.R
\name{methScores}
\alias{methScores}
\title{methScores
Wrapper function to calculate different methylation scores.}
\usage{
methScores(
  data,
  score = c("sampEn", "shannonEn"),
  axis = c("w", "h", "reads"),
  posCol = "pos",
  seqCol = "chr",
  binMode = c("custom", "tiled", "fixed"),
  regions = NULL,
  startCol = "start",
  endCol = "end",
  binSize = 1000,
  getBg = FALSE,
  ...
)
}
\arguments{
\item{data}{methylation data, can be either a GRanges object with the methylated levels
in the meta columns, a BSseq object, or a data.table/data.frame with
the methylation levels of a sample in the columns, a position (posCols) and sequence name column (seqCol).}

\item{score}{score function to use. Scores currently implemented are sampEn and shannonEn}

\item{axis}{either "w" or "h". The axis "w" specifies the calculation of the scores across the (CpG) sequence of
a single cell/sample. The axis "h" specifies calculation of the scores across cells for a position/region of interest.}

\item{posCol}{Column specifying the position of a CpG if input is a data.frame/data.table.}

\item{seqCol}{Column specifying the chromosome name of a CpG if input is a data.frame/data.table.}

\item{binMode}{Mode for binning the sequence. Mode "costum" if regions of interest are provided via
the regions argument. Mode "tiled" if the sequence should be split in tiled bins based on the positions.
Mode "fixed" if the sequence should be split up in bins containing a fixed number of CpGs.
Modes fixed and tiled differ in the sense that tiled bins the sequences in equally-sized stretches
while fixed ensures that the bins contain the same number of CpGs.}

\item{regions}{If binMode="costum" GRanges object which contains the regions of interest}

\item{startCol}{Name of the startCol (will get removed)}

\item{binSize}{Size of the bins, either the length of the stretches
in nucleotides (binMode="tiled") or in CpGs (binMode="fixed").}

\item{...}{args specific for the score function used}

\item{endCols}{Name of the endCol (will get removed)}
}
\value{
If data.table/data.frame or GRanges objects are provided as an input a data.table with the scores
for the bins across the axis of interest are provided. If a BSseq object is provided the scores get added to the
colData (if axis="w") or the metadata (if axis="h").
}
\description{
Wrapper function to calculate different entropy scores for different dimensions
of methylation data (i.e. inter-cell/height axis (h) or intra-cell/width axis (w)).
}
\author{
: Emanuel Sonder
}
