% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{sampEn}
\alias{sampEn}
\title{Sample Entropy function}
\usage{
sampEn(
  data,
  cols = NULL,
  block = NULL,
  pos = NULL,
  weighDist = FALSE,
  m = 2,
  naMode = c("keep", "remove"),
  r = NULL,
  measure = "maximum",
  nCores = 1,
  getBg = FALSE,
  longFormat = FALSE,
  ...
)
}
\arguments{
\item{data}{data for which to calculate Sample Entropies. Format expected is
a data.table with the numerical data for which to calculate the Sample Entropies
in the columns.}

\item{cols}{vector with names of the columns for which to calculate the Sample Entropy.}

\item{block}{If the Sample Entropies should be calculated for subseries of the
original data, block is the name of the column(s) which indicate the subseries.}

\item{pos}{Numerical column which gives the positions of the elements (not necessarily needed, might be removed)}

\item{m}{Length of the subseries/templates}

\item{naMode}{How to treat NAs. Mode "remove" all NAs get removed from the series and Sample Entropy
is calculated for the series with removed NAs. Mode "keep" subseries/templates get constructed such that no
subseries/templates is intersected by a NA.
(KeepSampEn, Dong et al. 2019, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7514754/)}

\item{r}{distance threshold (tolerance) for which subseries/templates are counted as matching pairs.
If r=NULL, r is set to 0, thus only exact matches are considered.
Makes sense if the data has only few states (e.g. binary series).}

\item{measure}{distance measure to be used, all distance measures of the dist function can be used:
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski.}

\item{nCores, }{number of cores to be used, parallelization happens across the cols.}
}
\value{
data.table with entropy scores for the specified columns and subseries
}
\description{
Implementation of Sample Entropy (Richman and Moorman, 2000,
https://journals.physiology.org/doi/full/10.1152/ajpheart.2000.278.6.h2039).
The complexity measure Sample Entropy is obtained by determining the negative
logarithm of the number of matching subseries/templates of length m+1 (A) and length m (B).
Matching subseries/templates are within a distance of r each other.
}
