% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{shannonEn}
\alias{shannonEn}
\title{Shannon Entropy function}
\usage{
shannonEn(
  data,
  cols = NULL,
  block = NULL,
  normalize = TRUE,
  discretize = TRUE,
  nCores = 1,
  getBg = FALSE,
  longFormat = FALSE,
  ...
)
}
\arguments{
\item{data}{data for which to calculate Shannon Entropies. Format expected is
a data.table with the numerical data for which to calculate the Shannon Entropies
in the columns.}

\item{cols}{vector with names of the columns for which to calculate the Sample Entropy.}

\item{block}{If the Shannon Entropies should be calculated for subseries of the
original data, block is the name of the column(s) which indicate the subseries.}

\item{normalize}{(TRUE/FALSE) If the shannon entropy should be normalized by the number of states.}

\item{discretize}{Should the values be discretized to the next integer}

\item{nCores, }{number of cores to be used, parallelization happens across the cols.}
}
\value{
data.table with entropy scores for the specified columns and subseries
}
\description{
Implementation of Shannon Entropy
(Shannon, 1948, http://people.math.harvard.edu/~ctm/home/text/others/shannon/entropy/entropy.pdf)
}
